Registration & Restricted pages

+++ base/views.py
@@ -1,7 +1,9 @@
 from django.shortcuts import render, redirect
-from django.contrib import messages
-from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models import Q
 from .models import Room,Topic
 from django.contrib.auth.models import User
@@ -17,8 +19,13 @@ from .forms import RoomForm
 # ]
 
 def loginView(request):
+
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
+
     if request.method == "POST":
-        username = request.POST.get("username")
+        username = request.POST.get("username").lower()
         password = request.POST.get("password")
 
         try:
@@ -33,9 +40,26 @@ def loginView(request):
             return redirect("home")
         else:
             messages.error(request, "Username or password are incorrect !")
-    context = {}
 
+    context = {"page" : page}
     return render(request, "base/auth.html", context)
+
+def registerView(request):
+
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occured during registeration")
+
+    return render(request, "base/auth.html", {"form" : form})
 
 def logoutView(request):
 
@@ -43,13 +67,15 @@ def logoutView(request):
     return redirect("login")
 
 def home(request):
+
     q = request.GET.get("q") if request.GET.get("q") != None else ""
+
     rooms = Room.objects.filter(
         Q(topic__title__icontains=q) |
         Q(title__icontains=q) |
         Q(description__icontains=q)
         )
-    rooms_count = rooms.count()
+    rooms_count = rooms.count() # Get query row count
     topics = Topic.objects.all()
     context = {
         "rooms" : rooms, 
@@ -63,6 +89,7 @@ def room(request, pk):
     context = {"room" : room}
     return render(request, "base/room.html", context)
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -73,9 +100,13 @@ def createRoom(request):
     context = {"form" : form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def editRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("Get the hell out of here!!")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -85,8 +116,12 @@ def editRoom(request, pk):
     context = {"form" : form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Get the hell out of here!!")
 
     if request.method == "POST": # Check if it's a POST request
         room.delete()
